[cc提交] refactor: 统一赋值目标解析 + Binder 拆分
